+++ plugins/gs/src/components/catalog/EntityKratixResourcesCard/index.ts
@@ -1,0 +1,1 @@
+export { EntityKratixResourcesCard } from './EntityKratixResourcesCard';