+++ plugins/gs/src/components/hooks/useResourceRequestStatusDetails.ts
@@ -1,0 +1,51 @@
+import {
+  StatusAborted,
+  StatusError,
+  StatusOK,
+} from '@backstage/core-components';
+import { useTheme } from '@material-ui/core';
+import CancelOutlinedIcon from '@material-ui/icons/CancelOutlined';
+import CheckCircleOutlinedIcon from '@material-ui/icons/CheckCircleOutlined';
+import HelpOutlinedIcon from '@material-ui/icons/HelpOutlineOutlined';
+import { toSentenceCase } from '../utils/helpers';
+import { ResourceRequestStatuses } from '@internal/plugin-gs-common';
+
+export function useResourceRequestStatusDetails(status: string) {
+  const theme = useTheme();
+
+  let statusIcon;
+  let icon;
+  let color;
+
+  switch (status) {
+    case ResourceRequestStatuses.Completed:
+      statusIcon = StatusOK;
+      icon = CheckCircleOutlinedIcon;
+      color = theme.palette.status.ok;
+      break;
+
+    case ResourceRequestStatuses.Failed:
+      statusIcon = StatusError;
+      icon = CancelOutlinedIcon;
+      color = theme.palette.status.error;
+      break;
+
+    case ResourceRequestStatuses.Unknown:
+      statusIcon = StatusAborted;
+      icon = HelpOutlinedIcon;
+      color = theme.palette.status.aborted;
+      break;
+
+    default:
+      statusIcon = StatusError;
+      icon = CancelOutlinedIcon;
+      color = theme.palette.status.error;
+  }
+
+  return {
+    statusIcon,
+    icon,
+    color,
+    label: toSentenceCase(status.replace(/-/g, ' ')),
+  };
+}