+++ plugins/gs/src/components/kratix/ResourceRequestStatus/index.ts
@@ -1,0 +1,1 @@
+export { ResourceRequestStatus } from './ResourceRequestStatus';